+++ lib/src/enums/calendar_background_type.dart
@@ -1,0 +1,9 @@
+/// A enum for the calendar background appearance.
+enum CalendarBackgroundType {
+  /// The provided color will be applied only.
+  plainColor,
+
+  /// The transparency to the provided color will be applied
+  /// using a background blur.
+  transparentAndBlured;
+}