+++ lib/src/enums/enums.dart
@@ -1,0 +1,1 @@
+export 'calendar_background_type.dart';