+++ lib/src/src.dart
@@ -1,7 +1,8 @@
 export 'animations/animations.dart';
 export 'calendar/calendar.dart';
 export 'config/config.dart';
-export 'config/ui_config.dart';
+export 'config/package_config.dart';
+export 'enums/enums.dart';
 export 'extensions/extensions.dart';
 export 'functions.dart';
 export 'utils/utils.dart';